// src/enums/saveStatus.enum.ts
export enum SaveStatusEnum{
   ADD_TASK= "ADD_TASK",
   ADD_SUBTASK = "ADD_SUBTASK",
   EDIT_TASK = "EDIT_TASK",
   EDIT_SUBTASK = "EDIT_SUBTASK",
   DELETE_TASK = "DELETE_TASK",
   DELETE_SUBTASK = "DELETE_SUBTASK",
   ADD_GLOBAL_COMMENT = "ADD_GLOBAL_COMMENT",
   ADD_COMMENT = "ADD_COMMENT",
   DEL_COMMENT = "DEL_COMMENT",
   ADD_COMMENT_SUBTASK = "ADD_COMMENT_SUBTASK"
}